Rename `api.errors` to `api.exception_handlers`
